requirements.txt
opencv-python~=4.8.0.76
numpy~=1.24.3
matplotlib~=3.7.2
rembg~=2.0.50
Pillow~=10.0.0
tensorflow~=2.13.0
keras~=2.13.1
ipython~=8.15.0
flask~=2.3.3